requirements.txt
anyio==3.6.2
async-timeout==4.0.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.1
google-auth==2.13.0
google-auth-oauthlib==0.7.0
gspread==5.6.2
h11==0.14.0
httplib2==0.21.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
motor==3.1.1
oauth2client==4.1.3
oauthlib==3.2.2
orjson==3.8.1
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.2
pyparsing==3.0.9
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.5
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rq==1.11.1
rsa==4.9
sentry-sdk==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
twilio==7.15.0
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
wrapt==1.14.1
